refactor(permission): 收敛权限面板控制逻辑
